alertmanager: only load cfg when writing cfg
Background
---
The operator can merge configuration when the
[`AlertmanagerConfigSelector`](https://github.com/prometheus-operator/prometheus-operator/blob/83fe36566f4e0894eb5ffcd2638a0f039a17bdeb/pkg/apis/monitoring/v1/types.go#L1547)
is defined. We are validating (using upstream alertmanager) and
parsing (using internal config) even if we don't need to load config in
memory since the user opted to manage the configuration via the secret.

This can cause problems where there is a missmatch between alertmanager
configuration and the operator internal amcfg struct, for example
prometheus-operator#4174.

Solution
---
Only validate the configuration from upstream `alertmanager` function,
and then only load the configuration in memory with the operators struct
when we need to write back to the secret.

As a cluster admin, I want to be able to pass arbitrary alertmanager
config via the secret, which goes unvalidated by the operator, so that I
can make use of latest alertmanager features without being concerned
about feature drift between the operator and alertmanager.