Avoid problematic race condition on statefulset recreation
We have observed an issue where the operator fails to update
a sts and must delete, recreate, a race condition can occur.

The delete is a foreground cascading delete which causes
the sts to remain visible via the API server for an unknown
amount of time.

We have observed for all resources, the operator is receiving a
multitude of update requests before the delete is finalized.

This is causing a hot-loop of delete, recreate cycles which in
some cases appear never to resolve without manual intervention
i.e. (deleting the sts in the background).

Since the foreground delete will add a deletionTimestamp to the
existing spec, we can check its existence to determine we should
not continue processing the update.

While this change does not guarantee zero downtime of the pods
owned by the sts, it should at least minimize it and not cause
the problem to exacerbate.

Co-authored-by: Simon Pasquier <spasquie@redhat.com>
Signed-off-by: Sunil Thaha <sthaha@redhat.com>